add enemy1

+++ GameEngine/enemy.h
@@ -17,12 +17,12 @@ public:
     //HIT_SCORE0 is the score you get when you hit this enemy
     //DESTROY_SCORE is the score you get when this enemy get destroyed
     Enemy(Point v,Point p,double angle0,HitPoint* hit_point0,
-          Graphic *graphic0,int health0,int HIT_SCORE0,int DESTROY_SCORE0);
+          Graphic *graphic0,double health0,int HIT_SCORE0,int DESTROY_SCORE0);
     //do damge and return score
-    int Hit(int damage);
+    int Hit(double damage);
     virtual void Fire(Game& my_game){}
 private:
-    int health;
+    double health;
     int HIT_SCORE;
     int DESTROY_SCORE;
 };
